+++ src/main/java/pl/plajer/villagedefense/creatures/v1_15_R1/VillagerBuster.java
@@ -18,10 +18,7 @@
 
 package pl.plajer.villagedefense.creatures.v1_15_R1;
 
-import java.util.Arrays;
-import java.util.logging.Level;
 
-import net.minecraft.server.v1_15_R1.DamageSource;
 import net.minecraft.server.v1_15_R1.EntityHuman;
 import net.minecraft.server.v1_15_R1.EntityIronGolem;
 import net.minecraft.server.v1_15_R1.EntityPlayer;
@@ -30,7 +27,6 @@ import net.minecraft.server.v1_15_R1.EntityZombie;
 import net.minecraft.server.v1_15_R1.GenericAttributes;
 import net.minecraft.server.v1_15_R1.PathfinderGoalBreakDoor;
 import net.minecraft.server.v1_15_R1.PathfinderGoalFloat;
-import net.minecraft.server.v1_15_R1.PathfinderGoalHurtByTarget;
 import net.minecraft.server.v1_15_R1.PathfinderGoalLookAtPlayer;
 import net.minecraft.server.v1_15_R1.PathfinderGoalMoveTowardsRestriction;
 import net.minecraft.server.v1_15_R1.PathfinderGoalNearestAttackableTarget;
@@ -38,32 +34,25 @@ import net.minecraft.server.v1_15_R1.PathfinderGoalRandomLookaround;
 import net.minecraft.server.v1_15_R1.PathfinderGoalZombieAttack;
 import net.minecraft.server.v1_15_R1.World;
 
-import org.bukkit.Bukkit;
-import org.bukkit.Material;
 import org.bukkit.craftbukkit.v1_15_R1.CraftWorld;
-import org.bukkit.entity.EntityType;
-import org.bukkit.entity.LivingEntity;
-import org.bukkit.entity.Player;
-import org.bukkit.event.entity.EntityDeathEvent;
-import org.bukkit.inventory.ItemStack;
 
 import pl.plajer.villagedefense.creatures.CreatureUtils;
-import pl.plajer.villagedefense.utils.Debugger;
 
 /**
  * Created by Tom on 15/08/2014.
  */
-public class PlayerBuster extends EntityZombie {
+public class VillagerBuster extends EntityZombie {
 
-  public PlayerBuster(org.bukkit.World world) {
+  public VillagerBuster(org.bukkit.World world) {
     this(((CraftWorld) world).getHandle());
   }
 
-  public PlayerBuster(World world) {
+  public VillagerBuster(World world) {
     super(world);
 
     GoalSelectorCleaner.clearSelectors(this);
     getNavigation().q().b(true);
+
 
     this.goalSelector.a(0, new PathfinderGoalFloat(this));
     this.goalSelector.a(1, new PathfinderGoalBreakDoor(this, enumDifficulty -> true));
@@ -71,14 +60,16 @@ public class PlayerBuster extends EntityZombie {
     this.goalSelector.a(2, new PathfinderGoalZombieAttack(this, CreatureUtils.getZombieSpeed(), false));
     this.goalSelector.a(4, new PathfinderGoalMoveTowardsRestriction(this, CreatureUtils.getZombieSpeed()));
     this.goalSelector.a(5, new PathfinderGoalBreakDoorFaster(this));
-    this.goalSelector.a(7, new PathfinderGoalLookAtPlayer(this, EntityPlayer.class, 8.0F)); // this one to look at human
+    this.goalSelector.a(7, new PathfinderGoalLookAtPlayer(this, EntityVillager.class, 8.0F)); // this one to look at human
     this.goalSelector.a(7, new PathfinderGoalRandomLookaround(this));
-    this.targetSelector.a(1, new PathfinderGoalHurtByTarget(this));
-    this.targetSelector.a(2, new PathfinderGoalNearestAttackableTarget<>(this, EntityPlayer.class, true));// this one to target
+    //this.targetSelector.a(1, new PathfinderGoalHurtByTarget(this));
+    this.targetSelector.a(2, new PathfinderGoalNearestAttackableTarget<>(this, EntityPlayer.class, false));
     this.targetSelector.a(2, new PathfinderGoalNearestAttackableTarget<>(this, EntityHuman.class, false));
-    this.targetSelector.a(2, new PathfinderGoalNearestAttackableTarget<>(this, EntityVillager.class, false));
+    this.targetSelector.a(2, new PathfinderGoalNearestAttackableTarget<>(this, EntityVillager.class, true));// this one to target
     this.targetSelector.a(2, new PathfinderGoalNearestAttackableTarget<>(this, EntityIronGolem.class, false));
-    this.setHealth(1);
+
+
+    this.setHealth(10);
     this.p(true);
   }
 
@@ -86,21 +77,5 @@ public class PlayerBuster extends EntityZombie {
   protected void initAttributes() {
     super.initAttributes();
     this.getAttributeInstance(GenericAttributes.FOLLOW_RANGE).setValue(200.0D);
-  }
-
-  @Override
-  public boolean damageEntity(DamageSource damagesource, float f) {
-    if (damagesource != null && damagesource.getEntity() != null && damagesource.getEntity().getBukkitEntity().getType() == EntityType.PLAYER && !CreatureUtils.getPlugin().getUserManager().getUser((Player) damagesource.getEntity().getBukkitEntity()).isSpectator()) {
-      Debugger.debug(Level.INFO,
-          "[PlayerBuster] Is player spectator? " + damagesource.getEntity().getBukkitEntity().getName() + " " + CreatureUtils.getPlugin().getUserManager().getUser((Player) damagesource.getEntity().getBukkitEntity()).isSpectator());
-      ItemStack[] itemStack = new ItemStack[] {new ItemStack(Material.ROTTEN_FLESH)};
-      Bukkit.getServer().getPluginManager().callEvent(new EntityDeathEvent((LivingEntity) this.getBukkitEntity(), Arrays.asList(itemStack), expToDrop));
-      getBukkitEntity().getWorld().spawnEntity(getBukkitEntity().getLocation(), EntityType.PRIMED_TNT);
-      this.die();
-      return true;
-    } else {
-      super.damageEntity(damagesource, f);
-      return false;
-    }
   }
 }